sampleField: extract useful utilities
and put them in their own files.
This file is for defining the observations, so the helpers should
live elsewhere.
