+++ python/pfs_instmodel/examples/sampleField.py
@@ -1,105 +1,6 @@
 import numpy as np
 from pfs_instmodel.schema.probes import PROBE
 
-def keepNth(input, N):
-    for i in input:
-        if i%N == 0:
-            yield i
-
-def keepOdd(input):
-    for i in input:
-        if i%2 == 1:
-            yield i
-
-def keepEven(input):
-    for i in input:
-        if i%2 == 0:
-            yield i
-
-class Fiber(object):
-    SCIENCE = 1
-    ENGINEERING = 2
-    BLANK = 3
-
-class Slit(object):
-    """  A temporary definition of a slit. To be replaced by a method in pfs_utils/fiberIds.
-
-    That said, this structure clarifies some of logic of the slit layout.
-    """
-
-    block1 = ([Fiber.ENGINEERING] +
-              42 * [Fiber.SCIENCE] +
-              [Fiber.BLANK])
-    block2 = ([Fiber.ENGINEERING] +
-              45 * [Fiber.SCIENCE] +
-              [Fiber.BLANK])
-    block3 = ([Fiber.ENGINEERING] +
-              [Fiber.BLANK] +
-              42 * [Fiber.SCIENCE] +
-              [Fiber.BLANK])
-    block4 = ([Fiber.ENGINEERING] +
-              42 * [Fiber.SCIENCE] +
-              [Fiber.ENGINEERING])
-    block5 = ([Fiber.ENGINEERING] +
-              45 * [Fiber.SCIENCE] +
-              [Fiber.ENGINEERING])
-    gap = 19*[Fiber.BLANK]
-
-    slit1 = (block1 +
-             block2 +
-             block3 +
-             block2 +
-             block3 +
-             block1 +
-             block4 +
-             gap +
-             block5 +
-             block1[::-1] +
-             block3[::-1] +
-             block1[::-1] +
-             block3[::-1] +
-             block2[::-1] +
-             block1[::-1])
-
-    # The second slit type is identical to the first but with
-    # three blanked-off science fibers.
-    slit2 = list(slit1)
-    for f in (280, 309, 359):
-        slit2[f-1] = Fiber.BLANK
-    slit2 = tuple(slit2)
-
-    slits = slit1, slit1, slit2, slit2
-
-    def __init__(self, slitId):
-        self.fiber0 = 1
-        scienceFibers = []
-        engineeringFibers = []
-
-        for i, fiberType in enumerate(self.slits[slitId]):
-            if fiberType is Fiber.SCIENCE:
-                scienceFibers.append(i)
-            elif fiberType is Fiber.ENGINEERING:
-                engineeringFibers.append(i)
-
-        self._scienceFibers = np.array(scienceFibers, dtype='i4')
-
-    @property
-    def scienceFibers(self):
-        return self._scienceFibers + self.fiber0
-    
-    def scienceFiberToSlitPos(self, scienceFiberNum):
-        """ Return the slit position for the given science fiber.
-        
-        This is only used for the engineering slits as LAM, which were 
-        sent defined in terms of the science fibers.
-        """
-
-        if scienceFiberNum < 0:
-            indexNum = scienceFiberNum + 300
-        else:
-            indexNum = scienceFiberNum + 299
-            
-        return self.scienceFibers[indexNum]
 
 Sm1Slit = Slit(1).scienceFibers
